shut down an MSVC warning.
git-svn-id: svn://svn.lyx.org/lyx/lyx-devel/trunk@21089 a592a061-630c-0410-9148-cb99ea01b6c8

+++ src/DepTable.cpp
@@ -59,7 +59,7 @@ void DepTable::insert(FileName const & f, bool upd)
 			LYXERR(Debug::DEPEND) << "done." << endl;
 			struct stat f_info;
 			stat(f.toFilesystemEncoding().c_str(), &f_info);
-			di.mtime_cur = f_info.st_mtime;
+			di.mtime_cur = long(f_info.st_mtime);
 		} else {
 			di.crc_cur = 0;
 			di.mtime_cur = 0;
